DjangoUI/requirements.txt
-i https://pypi.org/simple

asgiref==3.3.1; python_version >= '3.5'

certifi==2020.11.8

cffi==1.14.4

chardet==3.0.4

cryptography==3.2.1

django==3.1.4

idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

msal==1.6.0

pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pyjwt[crypto]==1.7.1

python-dotenv==0.15.0

pytz==2020.4

requests==2.25.0

six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sqlparse==0.4.1; python_version >= '3.5'

urllib3==1.26.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

